APPENDUID feature
http://tools.ietf.org/html/rfc4315
-----
[Page 2:]
The following response codes are extensions to the response codes
defined in [IMAP] section 7.1.

APPENDUID

Followed by the UIDVALIDITY of the destination mailbox and the UID
assigned to the appended message in the destination mailbox,
indicates that the message has been appended to the destination
mailbox with that UID.
-----

The SESSION_OK macro has been refactored to SESSION_OK_COMMON.
This is to notice if upstream changes the macro (resolving conflicts is better than duplicating code).
The refactored macro enables response codes.
SESSION_OK calls SESSION_OK_COMMON, so that it will behave the same as previously.
SESSION_OK_WITH_RESP_CODE may be used to send OK response with response code.
_ic_append_enter() has been changed:
 Creates a GString buffer.
 Copies the APPENDUID response code's contents to it with g_string_printf().
 Calls SESSION_OK_WITH_RESP_CODE with the buffer's contents.

Test changes:
 MyIMAP extends Net::IMAP.
 Greeting test is now improved.
 Testcase for append (regressions, answer format and data validity).

(ported and extended based on commit 34206bbc83e0a3b0213b698d35f81df090778b4b)