COPYUID feature (Feature #3789)
http://tools.ietf.org/html/rfc4315
-----
[Page 2-3:]
The following response codes are extensions to the response codes
defined in [IMAP] section 7.1.

COPYUID

Followed by the UIDVALIDITY of the destination mailbox, a UID set
containing the UIDs of the message(s) in the source mailbox that
were copied to the destination mailbox and containing the UIDs
assigned to the copied message(s) in the destination mailbox,
indicates that the message(s) have been copied to the destination
mailbox with the stated UID(s).

The source UID set is in the order the message(s) were copied; the
destination UID set corresponds to the source UID set and is in
the same order.
-----

ImapSession struct extended with new_ids list.
_do_copy() saves new_ids into this list.
_ic_copy_enter():
 retrieves old and new uids.
 Converts them to string.
 Creates and fills response code buffer.
 Calls SESSION_OK_WITH_RESP_CODE with the buffer's contents.

Test changes:
 ActiveSupport is included.
 Testcase for copy (regressions, answer format and data validity).

(ported and extended based on commit e374300699d18240986fd25ad701bbb4f74afb73)