IMAP: fix hierarchy elements masking out real folders in list response (Bug #7198)
The bug was reported to upstream:
 http://www.dbmail.org/mantis/view.php?id=978

The \Noselect flag was appearing unexpectedly in specific cases in list command responses.
This was reproducible for '%' wildcard search.
It was caused by the hierarchy elements found earlier and masking out real folders.
 If subfolders appear first in mailboxes_by_regex()'s response, then the
 partial hierarchy name is found before the real folder.
 All partial hierarchy elements receive \Noselect flag, so showing the
 name from the partial match will include this.
We could change the ordering of returned folders in mailboxes_by_regex(),
 but that would be only a workaround.
The real folders should have a higher priority than the hierarchy elements.
Store the found real folders and hierarchy elements into two separate trees
 before writing them out.
Insert folders from the lower priority tree (the hierarchy elements) to the
 higher priority one, except those that are already there.
 This way they won't mask real folders out.
Iterate the higher priority tree, and output folders to the socket.
It was necessary to insert new functions for the two iterations.
Several parts of the code have been refactored.
A new MailBox state info pointer needs to be reserved and stored into these
 trees.
Freeing can happen at different points.

Tests extended.