Imap: Folder in Cyrillic doesn't show \Haschildren flag (#987)
This was a string-processing error in dm_db.c/mailbox_match_new().
mailbox_match_new() is called at least from two places with a folder name pattern
 that contains '%' (this is for matching subfolders).
In the old version it was not excluding '%' from replacing it (in specific cases)
 with '_' ('&BBoEOARXBDI-/%' => '&BBoEOARXBDI___').
The '%' sign was replaced with '_' in the 'sensitive' part, and that is
 included in the SQL just when there is a utf7-imap-encoded part in the name.
 That's why the bug was reproducible just for names including parts in utf7-imap encoding.
This way there wasn't any match when counting subfolders of eg. Cyrillic folders.

Also a smaller fix was applied to properly change from verbatim to non-verbatim state.
The '-' sign - closing modified BASE64 data - is still part of that data.

Tests extended.